LayoutTests/fast/forms/date/date-pseudo-classes-expected.txt

Test if an input field matches to various pseudo classes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS matchedNodesContainId(container, ":enabled", "enabled") is true
PASS matchedNodesContainId(container, ":valid", "enabled") is true
PASS matchedNodesContainId(container, ":optional", "enabled") is true
PASS matchedNodesContainId(container, ":read-write", "enabled") is true
PASS matchedNodesContainId(container, ":disabled", "disabled") is true
PASS matchedNodesContainId(container, ":invalid", "invalid") is true
PASS matchedNodesContainId(container, ":in-range", "inrange") is true
PASS matchedNodesContainId(container, ":in-range", "badvalue") is true
PASS matchedNodesContainId(container, ":out-of-range", "outofrange") is true
PASS matchedNodesContainId(container, ":read-only", "readonly") is true
PASS matchedNodesContainId(container, ":required", "required") is true
PASS matchedNodesContainId(container, ":out-of-range", "outofrange") is false
PASS successfullyParsed is true

TEST COMPLETE
